Clean GA sampler
