+++ Optuna/Sampler/NSGAIISampler.cs
@@ -5,7 +5,7 @@ namespace Optuna.Sampler
     ///  <summary>
     /// https://optuna.readthedocs.io/en/stable/reference/generated/optuna.samplers.NSGAIISampler.html
     /// </summary>
-    public class NSGAIISampler : SamplerBase
+    public class NSGAIISampler : GASamplerBase
     {
         public double? MutationProb { get; set; }
         public int PopulationSize { get; set; } = 50;
@@ -24,29 +24,6 @@ namespace Optuna.Sampler
                 crossover: SetCrossover(optuna, Crossover),
                 constraints_func: hasConstraints ? ConstraintFunc() : null
             );
-        }
-
-        protected static dynamic SetCrossover(dynamic optuna, string crossover)
-        {
-            switch (crossover)
-            {
-                case "Uniform":
-                    return optuna.samplers.nsgaii.UniformCrossover();
-                case "BLXAlpha":
-                    return optuna.samplers.nsgaii.BLXAlphaCrossover();
-                case "SPX":
-                    return optuna.samplers.nsgaii.SPXCrossover();
-                case "SBX":
-                    return optuna.samplers.nsgaii.SBXCrossover();
-                case "VSBX":
-                    return optuna.samplers.nsgaii.VSBXCrossover();
-                case "UNDX":
-                    return optuna.samplers.nsgaii.UNDXCrossover();
-                case "":
-                    return null;
-                default:
-                    throw new ArgumentException("Unexpected crossover setting");
-            }
         }
     }
 }
